Moving config class to another file
