+++ lib/soter/config.rb
@@ -1,0 +1,17 @@
+module Soter
+  class Config < Struct.new(:fork, :logfile, :logger, :host, :port, :db,
+                            :workers)
+
+    def queue_settings
+      { 
+        host:       self.host,   
+        port:       self.port,
+        database:   self.db, 
+        collection: "mongo_queue",
+        timeout:    300,
+        attempts:   3
+      }
+    end
+
+  end
+end
